Move renderer/* files into sub-folders
