+++ src/mbgl/renderer/buckets/fill_bucket.cpp
@@ -1,9 +1,9 @@
-#include <mbgl/renderer/fill_bucket.hpp>
+#include <mbgl/renderer/buckets/fill_bucket.hpp>
 #include <mbgl/renderer/painter.hpp>
 #include <mbgl/programs/fill_program.hpp>
 #include <mbgl/renderer/bucket_parameters.hpp>
 #include <mbgl/style/layers/fill_layer_impl.hpp>
-#include <mbgl/renderer/render_fill_layer.hpp>
+#include <mbgl/renderer/layers/render_fill_layer.hpp>
 #include <mbgl/util/math.hpp>
 
 #include <mapbox/earcut.hpp>
