+++ src/mbgl/renderer/buckets/line_bucket.hpp
