+++ src/mbgl/renderer/buckets/raster_bucket.cpp
@@ -1,5 +1,5 @@
-#include <mbgl/renderer/raster_bucket.hpp>
-#include <mbgl/renderer/render_raster_layer.hpp>
+#include <mbgl/renderer/buckets/raster_bucket.hpp>
+#include <mbgl/renderer/layers/render_raster_layer.hpp>
 #include <mbgl/programs/raster_program.hpp>
 #include <mbgl/renderer/painter.hpp>
 #include <mbgl/gl/context.hpp>
