+++ src/mbgl/renderer/buckets/raster_bucket.hpp
