+++ src/mbgl/renderer/buckets/symbol_bucket.hpp
