+++ src/mbgl/renderer/layers/render_background_layer.hpp
