+++ src/mbgl/renderer/layers/render_raster_layer.hpp
