+++ src/mbgl/renderer/layers/render_symbol_layer.hpp
