+++ src/mbgl/renderer/painters/painter_clipping.cpp
