+++ src/mbgl/tile/raster_tile_worker.cpp
@@ -1,6 +1,6 @@
 #include <mbgl/tile/raster_tile_worker.hpp>
 #include <mbgl/tile/raster_tile.hpp>
-#include <mbgl/renderer/raster_bucket.hpp>
+#include <mbgl/renderer/buckets/raster_bucket.hpp>
 #include <mbgl/actor/actor.hpp>
 #include <mbgl/util/premultiply.hpp>
